+++ OneBlink.SDK/models/FormStoreAppForm.cs
@@ -1,0 +1,15 @@
+using System.Collections.Generic;
+namespace OneBlink.SDK.Model
+{
+    public class FormStoreForm
+    {
+        public long formId
+        {
+            get; set;
+        }
+        public List<string> groups
+        {
+            get; set;
+        }
+    }
+}